+++ lib/reduced_streambuilder.dart
@@ -1,5 +1,5 @@
 // reduced_streambuilder.dart;
 
 export 'src/inherited_widgets.dart';
-export 'src/streambuilder_reducible.dart';
-export 'src/streambuilder_wrapper.dart';
+export 'src/streambuilder_store.dart';
+export 'src/streambuilder_widgets.dart';