+++ test/reduced_streambuilder_test.dart
@@ -9,42 +9,19 @@ class Incrementer extends Reducer<int> {
 }
 
 void main() {
-  test('ReducibleBloc state 0', () {
+  test('Store state 0', () {
     final objectUnderTest = Store(0);
     expect(objectUnderTest.state, 0);
   });
 
-  test('ReducibleBloc state 1', () {
+  test('Store state 1', () {
     final objectUnderTest = Store(1);
     expect(objectUnderTest.state, 1);
   });
 
-  test('ReducibleBloc reduce', () async {
+  test('Store reduce', () async {
     final objectUnderTest = Store(0);
     objectUnderTest.reduce(Incrementer());
     expect(objectUnderTest.state, 1);
-  });
-
-  test('wrapWithProvider', () {
-    const child = SizedBox();
-    final objectUnderTest = wrapWithProvider(
-      initialState: 0,
-      child: child,
-    );
-    expect(objectUnderTest, isA<StatefulInheritedValueWidget<Store<int>>>());
-    final provider =
-        objectUnderTest as StatefulInheritedValueWidget<Store<int>>;
-    expect(provider.child, child);
-    final store = provider.value;
-    expect(store.state, 0);
-  });
-
-  test('wrapWithConsumer', () {
-    const child = SizedBox();
-    final objectUnderTest = wrapWithConsumer<int, int>(
-      builder: ({Key? key, required int props}) => child,
-      transformer: (reducible) => 1,
-    );
-    expect(objectUnderTest, isA<Builder>());
   });
 }