+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.1.0
 
 initial release.
+
+## 0.2.1 
+
+* replace function wrapWithProvider with widget class ReducedProvider
+* replace function wrapWithConsumer with widget class ReducedConsumer
+* replace function wrapWithScope with widget class ReducedScope
+* use reduced 0.2.1 from pub.dev